Set new selector session when creating nodegroup
Fixes https://github.com/weaveworks/eksctl/issues/3982 in which the
selector was not set on the nodegroup service object. This lead to a
panic when expanding instance selector options.

Integration test updated to use this codepath.

This commit also tidies up and fixes some errors in the nodegroup
create tests.